Simplify API docs plaintext build
