+++ astro/src/pages/[...lang]/api/latest.md.ts
@@ -1,15 +1,44 @@
 export const prerender = true;
 /**
- * AST-based plaintext rendering of the API Reference landing page.
+ * Plaintext rendering of the API Reference landing page.
  *
- * Equivalent to `latest.md.ts`, but composes the page from Markdoc nodes:
- * a heading, an intro paragraph, and a bullet list of category links.
+ * Composes the page from Markdoc nodes — a heading, an intro paragraph, and a
+ * bullet list of category links — then emits markdown via `buildMarkdocStr`.
+ * Mirrors the HTML landing page in `latest/index.astro`.
  */
 
+import type { Node as MarkdocNode } from "@markdoc/markdoc";
 import type { APIRoute, GetStaticPaths } from "astro";
+import type { ApiCategoryStub } from "@lib/api/schemas/views";
 import { getCategoryStubsView } from "@lib/api/viewsBuilder";
-import { LOCALES, parseLangParam } from "@lib/i18n/locale";
-import { apiLandingBody } from "@lib/plaintext/pages/apiPageBodies";
+import type { Locale } from "@lib/i18n/locale";
+import { LOCALES, localizedHref, parseLangParam } from "@lib/i18n/locale";
+import {
+  buildMarkdocStr,
+  heading,
+  inline,
+  link,
+  list,
+  listItem,
+  paragraphFromText,
+} from "@lib/plaintext/helpers";
+
+function apiLandingBody(categories: ApiCategoryStub[], lang: Locale): string {
+  const items = categories.map((cat) => {
+    const href = localizedHref(lang, `/api/latest/${cat.slug}/`);
+    return listItem([inline([link(href, cat.name)])]);
+  });
+
+  const contents: MarkdocNode[] = [
+    heading(1, "API Reference"),
+    paragraphFromText(
+      "Welcome to the Datadog API Reference. Select a category to get started.",
+    ),
+    list("unordered", items),
+  ];
+
+  return buildMarkdocStr(contents);
+}
 
 export const getStaticPaths: GetStaticPaths = () => {
   return LOCALES.map((lang) => ({
