+++ astro/src/pages/[...lang]/api/latest/[category]/[operation].md.ts
@@ -1,17 +1,30 @@
 export const prerender = true;
 /**
- * AST-based plaintext rendering of each endpoint page.
+ * Plaintext rendering of each endpoint page.
  *
- * Equivalent to `[operation].md.ts`, but builds a Markdoc AST and runs it
- * through `format()` rather than concatenating strings. The output is the
- * same shape — `# Summary` then per-variant `## v{N} (latest?)` sections —
- * but structure (tables, tabs, alerts, fences) is described as nodes.
+ * Builds the page as Markdoc nodes — `# Summary`, then a `## v{N} (latest?)`
+ * section per variant — and emits markdown via `buildMarkdocStr`. Structure
+ * (tables, tabs, alerts, fences) is described as nodes rather than concatenated
+ * strings. Mirrors the HTML endpoint page in `[operation].astro`.
  */
 
+import type { Node as MarkdocNode } from "@markdoc/markdoc";
 import type { APIRoute, GetStaticPaths } from "astro";
+import type { ApiOperationView } from "@lib/api/schemas/views";
 import { getCategoriesView, getOperationView } from "@lib/api/viewsBuilder";
 import { LOCALES, parseLangParam } from "@lib/i18n/locale";
-import { apiOperationBody } from "@lib/plaintext/pages/apiPageBodies";
+import { apiEndpointNodes } from "@components/ApiEndpoint/plaintext/ApiEndpoint";
+import { buildMarkdocStr, heading } from "@lib/plaintext/helpers";
+
+function apiOperationBody(operation: ApiOperationView): string {
+  const contents: MarkdocNode[] = [heading(1, operation.summary)];
+  for (const [i, variant] of operation.variants.entries()) {
+    const label = i === 0 ? `${variant.version} (latest)` : variant.version;
+    contents.push(heading(2, label));
+    contents.push(...apiEndpointNodes(variant));
+  }
+  return buildMarkdocStr(contents);
+}
 
 export const getStaticPaths: GetStaticPaths = async () => {
   const paths: ReturnType<GetStaticPaths> = [];
